Prototype Implementation of Weapon Prefabs
An implementation of the Weapon class designed to be used for the prototype. A Test scene Has been created. If you run the scene, An object should spawn a sample melee and ranged weapons. Controls are as follows:
A: Melee Attack
S: Range Attack
D: Melee Reload (Should do nothing)
F: Range Reload.

Additionally, there are sample prefabs if anyone wants to test the implementation,