Hit Detection On Weapons
Re organized the weapons folders into the prefab, scrip, and weapon testing assets folder. Added OnTriggerEnter2D script for the weapon hit boxes that call the Enemy ChangeHP(float) method. The EnemyZombie.Die() script is throwing an error. It does not appear to be implemented. it is being called from Enemy.ChangeHP().